Update onboarding flow to match new designs
