+++ client/src/css/wizard/Onboard2.css
@@ -1,10 +1,3 @@
-.personal-information {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-left: 10%;
-}
-
 .personal-information .page-text {
   width: 100%;
   font-size: 24px;
@@ -12,31 +5,26 @@
   text-align: center;
 }
 
-.personal-information-wrapper {
-  display: flex;
-  justify-content: left;
-  width: 90%;
-  margin-top: 20px;
-}
-
 .personal-information .personal-information-wrapper .svg {
   width: 100%;
   margin-top: 50px;
 }
 
-.personal-information .personal-information-wrapper .personal-information-form {
-  display: flex;
-  width: 70%;
-  justify-content: center;
+.personal-information-wrapper .personal-information-svg {
+  padding: 16px;
 }
 
-.personal-information
-  .personal-information-wrapper
-  .personal-information-form
-  .section {
-  margin: 20px 50px 20px 50px;
-  width: 70%;
+.personal-information-wrapper {
+  padding: 16px;
   display: flex;
-  flex-direction: column;
-  align-items: center;
+}
+
+.personal-information-wrapper .personal-information-form {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+}
+
+.personal-information-wrapper .personal-information-form .section {
+  padding: 8px;
+  text-align: left;
 }
