--- client/src/css/wizard/Onboard7.css
@@ -1,40 +1,0 @@
-.onboard7-wrapper {
-  width: 60%;
-  float: right;
-}
-
-.onboard7-wrapper .page-svg {
-  position: absolute;
-  left: 10vh;
-  bottom: 20vh;
-  width: 35%;
-}
-
-.onboard7-wrapper .onboard7-content {
-  margin-top: 10vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.onboard7-wrapper .onboard7-content .input-wrapper {
-  position: relative;
-  text-align: left;
-  width: 40%;
-  margin-bottom: 20px;
-}
-
-.onboard7-wrapper .onboard7-content .input-wrapper .input {
-  width: 100%;
-  /* margin-top: 0px; */
-  border-radius: 1rem;
-  background-color: rgb(61, 79, 145, 0.25);
-  padding: 10px;
-}
-
-.onboard7-wrapper .onboard7-content .input-wrapper .label {
-  /* position: absolute; */
-  font-size: 18px;
-  font-weight: 600;
-  margin-bottom: 7px;
-}
