+++ client/src/css/wizard/WizardListTitle.css
@@ -7,9 +7,9 @@
 .list-title {
   background-color: rgb(61, 79, 145, 0.55);
   color: white;
-  border-radius: 1.2rem;
+  border-radius: 10px;
   font-size: 24px;
-  padding: 12px 28px 12px 28px;
+  padding: 8px 24px 8px 24px;
   margin-bottom: 15px;
   width: fit-content;
   height: fit-content;
