+++ client/src/css/wizard/WizardSvg.css
@@ -1,4 +1,4 @@
-@media only screen and (max-width: 600px) {
+@media only screen and (max-width: 800px) {
   .page-svg {
     display: none;
   }
