+++ client/src/pages/wizard/Onboard3.tsx
@@ -1,94 +1,138 @@
 import React, {
   Dispatch,
-  SetStateAction,
   FC,
   ReactElement,
-  MouseEvent,
+  SetStateAction,
+  FormEvent,
 } from 'react';
+import { Checkbox, CheckboxProps, Form } from 'semantic-ui-react';
 
 import WizardSvg from '../../components/WizardSvg';
 import { handleSelectGroupArray } from '../../utils/helpers';
-import { racesEnum } from '../../utils/enums';
+import { interestsEnum } from '../../utils/enums';
 import '../../css/wizard/Onboard3.css';
-import WizardSelectButton from '../../components/WizardSelectButton/WizardSelectButton';
-import WizardListTitle from '../../components/WizardListTitle';
+import WizardStar from '../../components/WizardStar';
 
 interface IProps {
-  races: Array<string>;
-  setRaces: Dispatch<SetStateAction<Array<string>>>;
+  currentTeams: Array<string>;
+  setCurrentTeams: Dispatch<SetStateAction<Array<string>>>;
+  interests: Array<string>;
+  setInterests: Dispatch<SetStateAction<Array<string>>>;
 }
 
 /**
- * Builds and controls the form management for Onboard3 of the Onboarding Wizard
+ * Builds and controls the form management for Onboard5 of the Onboarding Wizard
  *
- * @param {Array<string>} races the selected races
- * @param {Dispatch<SetStateAction<Array<string>>>} setRaces React setter to update the selected races
+ * @param {Array<string>} currentTeams the current teams that the user is interested in
+ * @param {Dispatch<SetStateAction<Array<string>>>} setCurrentTeams React setter function to control
+ *                                                                  current Teams state variable
+ * @param {string} interests the interests of the user
+ * @param {Dispatch<SetStateAction<Array<string>>} setInterests React setter function to update user interests
  */
-const Onboard3: FC<IProps> = ({ races, setRaces }): ReactElement => {
+const Onboard3: FC<IProps> = ({
+  currentTeams,
+  setCurrentTeams,
+  interests,
+  setInterests,
+}): ReactElement => {
   /**
-   * Adds selected races to the races form array if it isn't already there, otherwise removes it
+   * Adds selected current team to the current teams form array if it isn't already there, otherwise removes it
    *
-   * @param e the mouse event from clicking one of the gender select optoins
+   * @param e the mouse event of clicking one of the current team select options
    */
-  const handleRaces = (e: MouseEvent<HTMLButtonElement>): void => {
-    handleSelectGroupArray(e, races, setRaces);
+  const handleCurrentTeams = (event: FormEvent<HTMLInputElement>): void => {
+    handleSelectGroupArray(event, currentTeams, setCurrentTeams);
   };
 
-  // All of the race buttons
-  const raceButtons = [
+  /**
+   * Adds selected interest to the interests form array if it isn't already there, otherwise removes it
+   *
+   * @param e the mouse event from clicking one of the interests select optoins
+   */
+  const handleInterests = (event: FormEvent<HTMLInputElement>): void => {
+    handleSelectGroupArray(event, interests, setInterests);
+  };
+
+  // All of the buttons to show for the current teams
+  const currentTeamsButtons = [
+    { value: 'Editing', color: '#A5C4F2' },
+    { value: 'Fact-checking', color: '#CFE7C4' },
+    { value: 'Illustration', color: '#BAB9E9' },
+    { value: 'Photography', color: '#D8ACE8' },
+    { value: 'Visuals', color: '#BFEBE0' },
+    { value: 'Writing', color: '#A9D3E5' },
+  ];
+
+  // All of the interests buttons to show
+  const interestsButtons = [
+    { display: 'Cannabis', value: interestsEnum.CANNABIS, color: '#CFE7C4' },
+    { display: 'Education', value: interestsEnum.EDUCATION, color: '#A9D3E5' },
     {
-      display: 'American Indian or Alaskan Native',
-      value: racesEnum.AMERICAN_INDIAN_OR_ALASKAN_NATIVE,
-      color: '#EF8B8B',
+      display: 'Food and Land',
+      value: interestsEnum.FOOD_AND_LAND,
+      color: '#BFEBE0',
+    },
+    { display: 'Fun', value: interestsEnum.FUN, color: '#F9B893' },
+    { display: 'Health', value: interestsEnum.HEALTH, color: '#F9B893' },
+    { display: 'Housing', value: interestsEnum.HOUSING, color: '#EF8B8B' },
+    {
+      display: 'Immigration',
+      value: interestsEnum.IMMIGRATION,
+      color: '#D8ACE8',
+    },
+    { display: 'Literature', value: interestsEnum.LIT, color: '#A5C4F2' },
+    { display: 'Music', value: interestsEnum.MUSIC, color: '#BFEBE0' },
+    { display: 'Nature', value: interestsEnum.NATURE, color: '#CFE7C4' },
+    { display: 'Politics', value: interestsEnum.POLITICS, color: '#A5C4F2' },
+    {
+      display: 'Stage and Screen',
+      value: interestsEnum.STAGE_AND_SCREEN,
+      color: '#D8ACE8',
     },
     {
-      display: 'Black or African American',
-      value: racesEnum.BLACK_OR_AFRICAN_AMERICAN,
-      color: '#A5C4F2',
+      display: 'Transportation',
+      value: interestsEnum.TRANSPORTATION,
+      color: '#F1D8B0',
     },
     {
-      display: 'Middle Eastern or North African',
-      value: racesEnum.MIDDLE_EASTERN_OR_NORTH_AFRICAN,
+      display: 'Visual Arts',
+      value: interestsEnum.VISUAL_ARTS,
       color: '#BAB9E9',
     },
-    {
-      display: 'Native Hawaiian or Pacific Islander',
-      value: racesEnum.NATIVE_HAWAIIAN_OR_PACIFIC_ISLANDER,
-      color: '#A9D3E5',
-    },
-    {
-      display: 'Latinx or Hispanic',
-      value: racesEnum.LATINX_OR_HISPANIC,
-      color: '#F9B893',
-    },
-    { display: 'White', value: racesEnum.WHITE, color: '#F1D8B0' },
-    { display: 'Asian', value: racesEnum.ASIAN, color: '#CFE7C4' },
-    { display: 'Other', value: racesEnum.OTHER, color: '#BFEBE0' },
   ];
 
   return (
-    <div className="races-wrapper">
-      <div className="page-text">
-        These fields are optional. If you feel comfortable answering, your
-        responses will not be used to evaulate your submission. Select all that
-        apply or leave it blank.
-      </div>
-      <div className="races-content-wrapper">
-        <WizardSvg page="onboard3" />
-        <WizardListTitle value="Race" />
-        <div className="race-button-wrapper">
-          {raceButtons.map((button) => (
-            <WizardSelectButton
-              key={button.value}
-              onClick={handleRaces}
-              buttonText={button.display}
-              value={button.value}
-              selectedArray={races}
-              color={button.color}
-              padding="5px 18px 5px 18px"
-              margin="20px 0px 20px 0px"
-            />
-          ))}
+    <div className="onboard5-wrapper">
+      <WizardSvg page="onboard3" />
+      <div className="onboard5-content">
+        <div className="section">
+          <div className="page-text">
+            <WizardStar />
+            <b>What do you want to do at the Weekly?</b>
+            <br />
+            Please limit to 1-2 teams that you're interested in.
+          </div>
+
+          <div className="select-group">
+            {currentTeamsButtons.map((button) => (
+              <Form.Field key={button.value} className="select-item">
+                <Checkbox label={button.value} onChange={handleCurrentTeams} />
+              </Form.Field>
+            ))}
+          </div>
+        </div>
+        <div className="section">
+          <div className="page-text">
+            <WizardStar />
+            <b>What topics are you interested in working on?</b>
+          </div>
+          <div className="select-group">
+            {interestsButtons.map((button) => (
+              <Form.Field key={button.value} className="select-item">
+                <Checkbox label={button.value} onChange={handleInterests} />
+              </Form.Field>
+            ))}
+          </div>
         </div>
       </div>
     </div>
