+++ client/src/pages/wizard/WizardWrapper.tsx
@@ -20,9 +20,6 @@ import Onboard2 from './Onboard2';
 import Onboard3 from './Onboard3';
 import Onboard4 from './Onboard4';
 import Onboard5 from './Onboard5';
-import Onboard6 from './Onboard6';
-import Onboard7 from './Onboard7';
-import Onboard8 from './Onboard8';
 import Compleition from './Completion';
 
 /**
@@ -35,9 +32,6 @@ enum WizardPage {
   ONBOARD_3 = 'ONBOARD_3',
   ONBOARD_4 = 'ONBOARD_4',
   ONBOARD_5 = 'ONBOARD_5',
-  ONBOARD_6 = 'ONBOARD_6',
-  ONBOARD_7 = 'ONBOARD_7',
-  ONBOARD_8 = 'ONBOARD_8',
   COMPLETION = 'COMPLETION',
 }
 
@@ -75,9 +69,7 @@ const WizardWrapper = (): ReactElement => {
     if (role === 'STAFF') {
       parsedPages = parsedPages.filter(
         (page) =>
-          page !== WizardPage.ONBOARD_4 &&
-          page !== WizardPage.ONBOARD_8 &&
-          page !== WizardPage.ONBOARD_5,
+          page !== WizardPage.ONBOARD_4 && page !== WizardPage.ONBOARD_5,
       );
     }
     setViewablePages(parsedPages);
@@ -192,7 +184,7 @@ const WizardWrapper = (): ReactElement => {
   useEffect(() => {
     if (
       scheduleConfirmed &&
-      page === WizardPage.ONBOARD_8.toString() &&
+      page === WizardPage.ONBOARD_5.toString() &&
       role === 'CONTRIBUTOR'
     ) {
       submitForm();
@@ -208,9 +200,9 @@ const WizardWrapper = (): ReactElement => {
    */
   const shouldShowNextBtn = (): boolean => {
     if (role === 'STAFF') {
-      return page !== WizardPage.ONBOARD_7.toString();
+      return page !== WizardPage.ONBOARD_3.toString();
     } else if (role === 'CONTRIBUTOR') {
-      return page !== WizardPage.ONBOARD_8.toString();
+      return page !== WizardPage.ONBOARD_5.toString();
     }
     return false;
   };
@@ -256,42 +248,28 @@ const WizardWrapper = (): ReactElement => {
               pronouns={pronouns}
               setGenders={setGenders}
               setPronouns={setPronouns}
+              races={races}
+              setRaces={setRaces}
             />
           )}
 
           {page === WizardPage.ONBOARD_3.toString() && (
-            <Onboard3 races={races} setRaces={setRaces} />
+            <Onboard3
+              currentTeams={currentTeams}
+              setCurrentTeams={setCurrentTeams}
+              interests={interests}
+              setInterests={setInterests}
+            />
           )}
-
           {page === WizardPage.ONBOARD_4.toString() && (
             <Onboard4
               reasonsForInvolvement={reasonForInvolvement}
               setReasonsForInvolvement={setReasonForInvolvement}
             />
           )}
+
           {page === WizardPage.ONBOARD_5.toString() && (
             <Onboard5
-              currentTeams={currentTeams}
-              setCurrentTeams={setCurrentTeams}
-            />
-          )}
-          {page === WizardPage.ONBOARD_6.toString() && (
-            <Onboard6 interests={interests} setInterests={setInterests} />
-          )}
-
-          {page === WizardPage.ONBOARD_7.toString() && (
-            <Onboard7
-              portfolio={portfolio}
-              linkedIn={linkedIn}
-              twitter={twitter}
-              setPortfolio={setPortfolio}
-              setLinkedIn={setLinkedIn}
-              setTwitter={setTwitter}
-            />
-          )}
-
-          {page === WizardPage.ONBOARD_8.toString() && (
-            <Onboard8
               isModalOpen={openModal}
               setScheduleConfirmed={setScheduleConfirmed}
               setModalOpen={setOpenModal}
